move choose type file to screens Folder
